third_party/blink/web_tests/svg/hittest/pointer-events-all2-expected.txt

Regression test for http://crbug.com/350338. This tests pointer-events:all on various shapes to make sure stroke:none and fill:none does not affect event processing. If this test passes, you will see "PASS" below.

PASS

(16, 36)
(54, 58)
(28, 77)
(139, 20)
(129, 41)
(149, 55)
(166, 70)
(128, 82)
(219, 19)
(242, 40)
(253, 64)
(318, 84)
(323, 47)
(324, 18)
(375, 64)
(426, 28)
(455, 44)
(445, 71)
(461, 85)
(497, 26)
(536, 95)
(560, 70)
(555, 54)
(595, 38)
(614, 16)
(569, 33)
(16, 136)
(54, 158)
(28, 177)
(139, 120)
(129, 141)
(149, 155)
(166, 170)
(128, 182)
(219, 119)
(242, 140)
(253, 164)
(318, 184)
(323, 147)
(324, 118)
(375, 164)
(426, 128)
(455, 144)
(445, 171)
(461, 185)
(497, 126)
(536, 195)
(560, 170)
(555, 154)
(595, 138)
(614, 116)
(569, 133)
